+++ src/artifactregistry/disallowPublicAccess.ts
@@ -6,23 +6,21 @@ export const disallowPublicAccess = {
 	enforcementLevel: 'advisory' as EnforcementLevel,
 	validateResource: (args: ResourceValidationArgs, reportViolation: ReportViolation) => {
 		const publicPrincipals = ['allUsers', 'allAuthenticatedUsers']
+		const message =
+			'Artifact Registry Repository should not be anonymously or publicly accessible.'
 		if (args.type === 'gcp:artifactregistry/repositoryIamBinding:RepositoryIamBinding') {
 			const members = args.props.members
 			if (members) {
 				members.forEach((member: string) => {
 					if (publicPrincipals.includes(member)) {
-						reportViolation(
-							'Artifact Registry Repository should not be anonymously or publicly accessible.'
-						)
+						reportViolation(message)
 					}
 				})
 			}
 		} else if (args.type === 'gcp:artifactregistry/repositoryIamMember:RepositoryIamMember') {
 			const member = args.props.member
 			if (member && publicPrincipals.includes(member)) {
-				reportViolation(
-					'Artifact Registry Repository should not be anonymously or publicly accessible.'
-				)
+				reportViolation(message)
 			}
 		}
 	}