+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cloud DNS policies (#30)
   - `disallowRsasha1Algorithm`
   - `requireDnssecStateOn`
+- Pub/Sub policies (#31)
+  - `requireSubscriptionDeadLetterTopic`
+  - `requireTopicCmek`
 
 ## [1.6.0] - 2024-01-21