+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.7.0] - 2024-01-22
+
 ### Added
 
 - Cloud DNS policies [#30](https://github.com/losisin/gcp-pac/pull/30)